Move Kit-related functions and classes into kitmodel.cpp
